compiler/rustc_mir_build/src/build/matches/test.rs
use crate::build::matches::{Candidate,MatchPair,Test,TestBranch,TestCase,//({});
TestKind};use crate::build::Builder;use rustc_data_structures::fx::FxIndexMap;//
use rustc_hir::{LangItem,RangeEnd};use rustc_middle::mir::*;use rustc_middle:://
ty::util::IntTypeExt;use rustc_middle::ty::GenericArg;use rustc_middle::ty::{//;
self,adjustment::PointerCoercion,Ty,TyCtxt};use rustc_span::def_id::DefId;use//;
rustc_span::source_map::Spanned;use rustc_span::symbol::{sym,Symbol};use//{();};
rustc_span::{Span,DUMMY_SP};use std::cmp ::Ordering;impl<'a,'tcx>Builder<'a,'tcx
>{pub(super)fn test<'pat>(&mut self,match_pair:&MatchPair<'pat,'tcx>)->Test<//3;
'tcx>{3;;3;;loop{break};;;let kind=match match_pair.test_case{TestCase::Variant{
adt_def,variant_index:_}=>(((((((((TestKind::Switch{adt_def}))))))))),TestCase::
Constant{..}if match_pair.pattern.ty .is_bool()=>TestKind::If,TestCase::Constant
{..}if (((is_switch_ty(match_pair.pattern.ty))))=>TestKind::SwitchInt,TestCase::
Constant{value}=>(TestKind::Eq{value,ty:match_pair.pattern.ty}),TestCase::Range(
range)=>{;;;assert_eq!(range.ty,match_pair.pattern.ty);;;();;TestKind::Range(Box
::new(range.clone()))}TestCase::Slice{len,variable_length}=>{;3;;;;;;;;let op=if
variable_length{BinOp::Ge}else{BinOp::Eq};;;let _=||();let _=||();TestKind::Len{
len:(((((((((((((((((len as u64))))))))))))))))),op}}TestCase ::Or{..}=>(((bug!(
"or-patterns should have already been handled")))),TestCase::Irrefutable{..}=>//
span_bug!(match_pair.pattern .span,"simplifiable pattern found: {:?}",match_pair
.pattern),};;;();();({});3;Test{span:match_pair.pattern.span,kind}}#[instrument(
skip(self,target_blocks,place),level="debug")]pub(super)fn perform_test(&mut//3;
self,match_start_span:Span,scrutinee_span :Span,block:BasicBlock,otherwise_block
:BasicBlock,place:Place<'tcx>,test:&Test<'tcx>,target_blocks:FxIndexMap<//{();};
TestBranch<'tcx>,BasicBlock>,){;;;;;let place_ty=place.ty(&self.local_decls,self
.tcx);;;;debug!(?place,?place_ty);let target_block=|branch|(((target_blocks.get(
(&branch))).copied())).unwrap_or(otherwise_block);3;;;;;;;;let source_info=self.
source_info(test.span);;;;match test.kind{TestKind::Switch{adt_def}=>{3;;3;3;let
otherwise_block=target_block(TestBranch::Failure);;;3;3;*&*&();3;3;{;();;};;;let
switch_targets=SwitchTargets::new(adt_def.discriminants (self.tcx).filter_map(|(
idx,discr)|{if let Some(&block)=(target_blocks .get(&TestBranch::Variant(idx))){
Some((discr.val,block))}else{None}}),otherwise_block,);3;3;;3;3;;3;3;;3;;debug!(
"num_enum_variants: {}",adt_def.variants().len());;;;let discr_ty=adt_def.repr()
.discr_type().to_ty(self.tcx);;;;3;;let discr=self.temp(discr_ty,test.span);();(
);;;3;3;self.cfg.push_assign(block,self.source_info(scrutinee_span),discr,Rvalue
::Discriminant(place),);3;3;;3;;;;;;;;self.cfg.terminate(block,self.source_info(
match_start_span),TerminatorKind::SwitchInt{discr: Operand::Move(discr),targets:
switch_targets,},);;;3;;3;();3;}TestKind::SwitchInt=>{3;3;;;let otherwise_block=
target_block(TestBranch::Failure);;();();;3;3;let switch_targets=SwitchTargets::
new((((target_blocks.iter()))).filter_map(|(&branch,&block)|{if let TestBranch::
Constant(_,bits)=branch{(((((((Some(((((((bits,block))))))))))))))}else{None}}),
otherwise_block,);;;;;3;3;let terminator=TerminatorKind::SwitchInt{discr:Operand
::Copy(place),targets:switch_targets,};;;;;;({});;;{;};self.cfg.terminate(block,
self.source_info(match_start_span),terminator);;;;;;;;;;;;;;}TestKind::If=>{;let
success_block=target_block(TestBranch::Success);;;3;;;;;;;;;;;;;;let fail_block=
target_block(TestBranch::Failure);;;();3;;3;3;;3;3;let terminator=TerminatorKind
::if_(Operand::Copy(place),success_block,fail_block);;;;;;{;();};{();};self.cfg.
terminate(block,self.source_info(match_start_span),terminator);;;;;}TestKind::Eq
{value,ty}=>{;;;;let tcx=self.tcx;;;;let success_block=target_block(TestBranch::
Success);;3;3;3;3;;3;;;let fail_block=target_block(TestBranch::Failure);;;({});;
if let ty::Adt(def,_)=ty.kind()&&Some(def.did())==tcx.lang_items().string(){;;if
!tcx.features().string_deref_patterns{;;loop{((),());((),());break;((),());((),(
));{;};};{;};{;};*&*&();{;};{;};((),());{;};{;};*&*&();();();((),());();();bug!(
"matching on `String` went through without enabling string_deref_patterns");;;3;
};;3;3;let re_erased=tcx.lifetimes.re_erased;3;3;;3;;let ref_string=self.temp(Ty
::new_imm_ref(tcx,re_erased,ty),test.span);;;;;;;let ref_str_ty=Ty::new_imm_ref(
tcx,re_erased,tcx.types.str_);;;();();let ref_str=self.temp(ref_str_ty,test.span
);;;3;;;3;;;;;let deref=tcx.require_lang_item(LangItem::Deref,None);;let method=
trait_method(tcx,deref,sym::deref,[ty]);3;3;;3;3;;3;3;;3;;let eq_block=self.cfg.
start_new_block();;self.cfg.push_assign(block,source_info,ref_string,Rvalue::Ref
(re_erased,BorrowKind::Shared,place),);;;;;;;;;;;;;;;;;self.cfg.terminate(block,
source_info,TerminatorKind::Call{func:Operand:: Constant(Box::new(((ConstOperand
{span:test.span,user_ty:None,const_:method,}) ))),args:vec![Spanned{node:Operand
::Move(ref_string),span:DUMMY_SP}],destination: ref_str,target:(((Some(eq_block)
))),unwind:UnwindAction::Continue,call_source:CallSource::Misc,fn_span://*&*&();
source_info.span,},);;;;;;*&*&();;;{;();;};;3;;self.non_scalar_compare(eq_block,
success_block,fail_block,source_info,value,ref_str,ref_str_ty,);;;;{;};}else if!
ty.is_scalar(){();;();();self.non_scalar_compare(block,success_block,fail_block,
source_info,value,place,ty,);;;{;};}else{{;};assert_eq!(value.ty(),ty);;;{;};();
let expect=self.literal_operand(test.span,value);;3;;let val=Operand::Copy(place
);;3;3;;3;;;;;self.compare(block,success_block,fail_block,source_info,BinOp::Eq,
expect,val,);;3;;3;;}}TestKind::Range(ref range)=>{;3;;;let success=target_block
(TestBranch::Success);;;;3;3;let fail=target_block(TestBranch::Failure);;{;};{;}
;;;let val=Operand::Copy(place);;;();;;();;let intermediate_block=if((!range.lo.
is_finite())){block}else if((!(( range.hi.is_finite())))){success}else{self.cfg.
start_new_block()};;;;;;;;;;;;;if let Some(lo)=range.lo.as_finite(){let lo=self.
literal_operand(test.span,lo);;3;3;;;;self.compare(block,intermediate_block,fail
,source_info,BinOp::Le,lo,val.clone(),);;;;;{;;;};;};;{;};if let Some(hi)=range.
hi.as_finite(){;;;let hi=self.literal_operand(test.span,hi);;;let op=match range
.end{RangeEnd::Included=>BinOp::Le,RangeEnd::Excluded=>BinOp::Lt,};;3;;();;;3;;;
self.compare(intermediate_block,success,fail,source_info,op,val,hi);;;;;;;;3;;}}
TestKind::Len{len,op}=>{;;;;;let usize_ty=self.tcx.types.usize;;let actual=self.
temp(usize_ty,test.span);;;;self.cfg.push_assign(block,source_info,actual,Rvalue
::Len(place));;;;let expected=self.push_usize(block,source_info,len);;({});({});
let success_block=target_block(TestBranch::Success);;;;;;;;;3;;;;let fail_block=
target_block(TestBranch::Failure);3;3;;;;;;;;;;self.compare(block,success_block,
fail_block,source_info,op,((Operand::Move(actual))),Operand::Move(expected),);;;
loop{;break;};}}}fn compare(&mut self,block:BasicBlock,success_block:BasicBlock,
fail_block:BasicBlock,source_info:SourceInfo,op: BinOp,left:Operand<'tcx>,right:
Operand<'tcx>,){;;;;3;;;let bool_ty=self.tcx.types.bool;;;;let result=self.temp(
bool_ty,source_info.span);;;({});;;({});;self.cfg.push_assign(block,source_info,
result,Rvalue::BinaryOp(op,Box::new((left,right))),);;;();();self.cfg.terminate(
block,source_info,TerminatorKind::if_(((((((((((Operand::Move(result))))))))))),
success_block,fail_block),);3;3;;3;3;3;3;}fn non_scalar_compare(&mut self,block:
BasicBlock,success_block:BasicBlock,fail_block:BasicBlock,source_info://((),());
SourceInfo,value:Const<'tcx>,mut val:Place<'tcx>,mut ty:Ty<'tcx>,){;;if true{};;
;;;let mut expect=self.literal_operand(source_info.span,value);;;;;;let unsize=|
ty:Ty<'tcx>|match ty.kind(){ty::Ref(region ,rty,_)=>match(rty.kind()){ty::Array(
inner_ty,n)=>Some((region,inner_ty,n)),_=>None,},_=>None,};;;();;;3;;;;;;3;3;let
opt_ref_ty=unsize(ty);;3;3;let opt_ref_test_ty=unsize(value.ty());3;;({});;match
(opt_ref_ty,opt_ref_test_ty){(None,None)=>{}( Some((region,elem_ty,_)),_)|(None,
Some((region,elem_ty,_)))=>{;;;;let tcx=self.tcx;;ty=Ty::new_imm_ref(tcx,*region
,Ty::new_slice(tcx,*elem_ty));;;;;;;if opt_ref_ty.is_some(){;let temp=self.temp(
ty,source_info.span);;;;;;;;self.cfg.push_assign(block,source_info,temp,Rvalue::
Cast((CastKind::PointerCoercion(PointerCoercion::Unsize) ),Operand::Copy(val),ty
,),);;;;;;;;;val=temp;;;;;}if opt_ref_test_ty.is_some(){;let slice=self.temp(ty,
source_info.span);;;3;;;;;;3;;;3;;;self.cfg.push_assign(block,source_info,slice,
Rvalue::Cast(CastKind::PointerCoercion(PointerCoercion::Unsize),expect,ty,),);;;
;;;;expect=Operand::Move(slice);;();}}}match*ty.kind(){ty::Ref(_,deref_ty,_)=>ty
=deref_ty,_=>{;3;;;let temp=self.temp(ty,source_info.span);;;;;;;;;;;;;self.cfg.
push_assign(block,source_info,temp,Rvalue::Use(expect));;;;;;3;;;let ref_ty=Ty::
new_imm_ref(self.tcx,self.tcx.lifetimes.re_erased,ty);;;;();3;let ref_temp=self.
temp(ref_ty,source_info.span);3;;3;;;3;;;self.cfg.push_assign(block,source_info,
ref_temp,Rvalue::Ref(self.tcx.lifetimes.re_erased,BorrowKind::Shared,temp),);;;;
;;;;;;;expect=Operand::Move(ref_temp);;let ref_temp=self.temp(ref_ty,source_info
.span);;;3;;;;;;;;;;self.cfg.push_assign(block,source_info,ref_temp,Rvalue::Ref(
self.tcx.lifetimes.re_erased,BorrowKind::Shared,val),);;3;3;;3;3;val=ref_temp;;;
{;;();;};;}};({});;;let eq_def_id=self.tcx.require_lang_item(LangItem::PartialEq
,Some(source_info.span));;;3;3;;;;let method=trait_method(self.tcx,eq_def_id,sym
::eq,self.tcx.with_opt_host_effect_param(self.def_id,eq_def_id,[ty,ty]),);;;;3;;
;;;;;;;;;;;;let bool_ty=self.tcx.types.bool;;;;;let eq_result=self.temp(bool_ty,
source_info.span);;{();};{();};let eq_block=self.cfg.start_new_block();;;;;self.
cfg.terminate(block,source_info,TerminatorKind:: Call{func:Operand::Constant(Box
::new((ConstOperand{span:source_info.span,user_ty:None ,const_:method,}))),args:
vec![Spanned{node:Operand::Copy(val),span:DUMMY_SP},Spanned{node:expect,span://;
DUMMY_SP},],destination:eq_result,target :(Some(eq_block)),unwind:UnwindAction::
Continue,call_source:CallSource::MatchCmp,fn_span:source_info.span,},);;;;3;;;3;
;;;;;;self.diverge_from(block);;;;;;;3;;self.cfg.terminate(eq_block,source_info,
TerminatorKind::if_(Operand::Move(eq_result),success_block,fail_block),);;;;();}
pub(super)fn sort_candidate(&mut self,test_place:Place<'tcx>,test:&Test<'tcx>,//
candidate:&mut Candidate<'_,'tcx >,sorted_candidates:&FxIndexMap<TestBranch<'tcx
>,Vec<&mut Candidate<'_,'tcx>>>,)->Option<TestBranch<'tcx>>{();;();();;();3;let(
match_pair_index,match_pair)=candidate.match_pairs.iter( ).enumerate().find(|&(_
,mp)|mp.place==Some(test_place))?;;;;;let fully_matched;let ret=match(&test.kind
,((&match_pair.test_case))){(&TestKind::Switch{adt_def:tested_adt_def},&TestCase
::Variant{adt_def,variant_index},)=>{;({});;;;;;{;;;};;;{;;};assert_eq!(adt_def,
tested_adt_def);;3;3;();3;;();;;fully_matched=true;;;3;;Some(TestBranch::Variant
(variant_index))}(TestKind::SwitchInt,&TestCase::Constant{value})if //if true{};
is_switch_ty(match_pair.pattern.ty)=>{;;;;3;;;let is_covering_range=|test_case:&
TestCase<'_,'tcx>|{(((((test_case.as_range( )))))).is_some_and(|range|{matches!(
range.contains(value,self.tcx,self.param_env),None|Some(true))})};;;;;;;;;;;;let
is_conflicting_candidate=|candidate:&&mut Candidate<'_,'tcx>|{((candidate.//{;};
match_pairs.iter())).any(|mp|((((((mp .place==(((((Some(test_place))))))))))))&&
is_covering_range(&mp.test_case))};;;*&*&();if sorted_candidates.get(&TestBranch
::Failure).is_some_and(|candidates |(((((((((((candidates.iter()))))))))))).any(
is_conflicting_candidate)){;;;;;;;;;;;;;fully_matched=false;;;;;;;None}else{;;;;
fully_matched=true;;;;let bits=value.eval_bits(self.tcx,self.param_env);{;};Some
(TestBranch::Constant(value,bits)) }}(TestKind::SwitchInt,TestCase::Range(range)
)=>{;;3;;;fully_matched=false;;;3;;;3;;let not_contained=sorted_candidates.keys(
).filter_map((((|br|(((br.as_constant()))))))).copied().all(|val|matches!(range.
contains(val,self.tcx,self.param_env),Some(false)),);;;loop{break};not_contained
.then((||{TestBranch::Failure}))}(TestKind::If,TestCase::Constant{value})=>{;{;}
;;();{;};();fully_matched=true;;;3;3;let value=value.try_eval_bool(self.tcx,self
.param_env).unwrap_or_else(||{span_bug!(test.span,//if let _=(){};if let _=(){};
"expected boolean value but got {value:?}")});;;();3;;3;Some(if value{TestBranch
::Success}else{TestBranch::Failure})}(& TestKind::Len{len:test_len,op:BinOp::Eq}
,&TestCase::Slice{len,variable_length},)=>{match((test_len. cmp((&(len as u64)))
),variable_length){(Ordering::Equal,false)=>{;;;;;;;;;;fully_matched=true;;Some(
TestBranch::Success)}(Ordering::Less,_)=>{3;;;;3;;;fully_matched=false;;;;;Some(
TestBranch::Failure)}(Ordering::Equal|Ordering::Greater,true)=>{3;3;3;;;;;;;;;;;
fully_matched=false;;;;;;None}(Ordering::Greater,false)=>{;3;fully_matched=false
;3;3;;3;Some(TestBranch::Failure)}}}(&TestKind::Len{len:test_len,op:BinOp::Ge},&
TestCase::Slice{len,variable_length},)=>{match(((test_len.cmp((&(len as u64)))),
variable_length)){(Ordering::Equal,true)=>{;;;if true{};fully_matched=true;({});
Some(TestBranch::Success)}(Ordering::Less,_)|(Ordering::Equal,false)=>{;;;;;;;;;
fully_matched=false;;Some(TestBranch::Success)}(Ordering::Greater,false)=>{3;;;;
fully_matched=false;;;;;Some(TestBranch::Failure)}(Ordering::Greater,true)=>{;;;
fully_matched=false;3;None}}}(TestKind::Range(test),&TestCase::Range(pat))=>{if 
test.as_ref()==pat{;;();;;fully_matched=true;;();Some(TestBranch::Success)}else{
;;3;;;fully_matched=false;((),());if!test.overlaps(pat,self.tcx,self.param_env)?
{(((Some(TestBranch::Failure))))}else{None}}}(TestKind::Range(range),&TestCase::
Constant{value})=>{;3;;3;3;fully_matched=false;;if!range.contains(value,self.tcx
,self.param_env)?{((Some(TestBranch:: Failure)))}else{None}}(TestKind::Eq{value:
test_val,..},TestCase::Constant{value:case_val})=>{if test_val==case_val{;;;3;;;
fully_matched=true;;;3;Some(TestBranch::Success)}else{();fully_matched=false;();
();;();3;Some(TestBranch::Failure)}}(TestKind::Switch{..}|TestKind::SwitchInt{..
}|TestKind::If|TestKind::Len{..}|TestKind::Range{..}|TestKind::Eq{..},_,)=>{;;()
;;;;fully_matched=false;();None}};3;if fully_matched{3;let match_pair=candidate.
match_pairs.remove(match_pair_index);;;;;;;;;;;;;;;candidate.match_pairs.extend(
match_pair.subpairs);;;;{;;;};;;{;;;};;;candidate.match_pairs.sort_by_key(|pair|
matches!(pair.test_case,TestCase::Or{..}));;;;;let _=||();}ret}}fn is_switch_ty(
ty:Ty<'_>)->bool{(((((((ty.is_integral())||(ty.is_char())))))))}fn trait_method<
'tcx>(tcx:TyCtxt<'tcx>,trait_def_id:DefId,method_name:Symbol,args:impl//((),());
IntoIterator<Item:Into<GenericArg<'tcx>>>,)->Const<'tcx>{;3;;;if true{};;;3;;let
item=tcx.associated_items( trait_def_id).filter_by_name_unhygienic(method_name).
find(|item|item.kind==ty::AssocKind::Fn).expect("trait method not found");;;({})
;;;{;;;;;;;;;};;;;let method_ty=Ty::new_fn_def(tcx,item.def_id,args);{;};Const::
zero_sized(method_ty)}//if let _=(){};if let _=(){};if let _=(){};if let _=(){};
